[FW][REF] (website_)knowledge: lint and cleanup tests
This commit contains some post-merge cleanup in knowledge tests before adding
some new tests

  * remove an unused common file, holding unused data;
  * some rephrasing and cleanup in comments;
  * move fields specific test into an "internal" file that holds tests for
    computed fields notably;

Task-2859616

X-Original-Commit: odoo/odoo@bad3ee069e8860d4259160e1222fd6a9fdaef562
