+++ addons/website_knowledge/tests/test_knowledge_security.py
@@ -2,19 +2,19 @@
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
 from odoo import exceptions
-from odoo.addons.knowledge.tests.common import KnowledgeArticlePermissionsCase
+from odoo.addons.knowledge.tests.common import KnowledgeCommonWData
 from odoo.tests.common import tagged, users
 from odoo.tools import mute_logger
 
 
 @tagged('knowledge_acl')
-class TestWKnowledgeSecurity(KnowledgeArticlePermissionsCase):
+class TestWKnowledgeSecurity(KnowledgeCommonWData):
 
     @mute_logger('odoo.addons.base.models.ir_model', 'odoo.addons.base.models.ir_rule')
     @users('user_public')
     def test_models_as_public(self):
         """ Test publish flag giving read access to articles """
-        article_shared = self.article_roots[2].with_env(self.env)
+        article_shared = self.article_shared.with_env(self.env)
         with self.assertRaises(exceptions.AccessError,
                                msg="ACLs: Internal permission 'none', not for public"):
             article_shared.body  # access body should trigger acls
@@ -30,7 +30,7 @@ class TestWKnowledgeSecurity(KnowledgeArticlePermissionsCase):
     @users('portal_test')
     def test_models_as_portal(self):
         """ Test publish flag giving read access to articles """
-        article_shared = self.article_roots[2].with_env(self.env)
+        article_shared = self.article_shared.with_env(self.env)
         with self.assertRaises(exceptions.AccessError,
                                msg="ACLs: Internal permission 'none', not for portal"):
             article_shared.body  # access body should trigger acls
@@ -48,7 +48,7 @@ class TestWKnowledgeSecurity(KnowledgeArticlePermissionsCase):
     @users('portal_test')
     def test_models_as_user(self):
         """ Test publish flag giving read access to articles """
-        article_hidden = self.article_read_contents[3].with_env(self.env)
+        article_hidden = self.private_children[0].with_env(self.env)
         with self.assertRaises(exceptions.AccessError,
                                msg="ACLs: 'none' internal permission"):
             article_hidden.body  # access body should trigger acls
